articles_api/requirements.txt
Django==2.2.1
djangorestframework==3.9.4
pytz==2019.1
#djangorestframework-simplejwt==4.30
gunicorn==19.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1